requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
blas=1.0=mkl
boltons=23.0.0=py310h06a4308_0
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
cffi=1.15.1=py310h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310h06a4308_0
conda=23.3.1=py310h06a4308_0
conda-content-trust=0.1.3=py310h06a4308_0
conda-package-handling=2.0.2=py310h06a4308_0
conda-package-streaming=0.7.0=py310h06a4308_0
cryptography=39.0.1=py310h9ce1e76_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py310h06a4308_0
glfw=2.5.9=pypi_0
gymnasium-robotics=1.2.0=pypi_0
idna=3.4=py310h06a4308_0
imageio=2.28.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46305
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libprotobuf=3.20.3=he621ea3_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py310h5eee18b_1
mkl_fft=1.3.6=py310h1128e8f_1
mkl_random=1.2.2=py310h1128e8f_1
mujoco=2.3.3=pypi_0
ncurses=6.4=h6a678d5_0
numpy=1.23.5=pypi_0
numpy-base=1.24.3=py310hb5e798b_1
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py310h06a4308_0
pathtools=0.1.2=pyhd3eb1b0_1
pillow=9.5.0=pypi_0
pip=23.0.1=py310h06a4308_0
pluggy=1.0.0=py310h06a4308_1
protobuf=3.20.3=py310h6a678d5_0
psutil=5.9.0=py310h5eee18b_0
pycosat=0.6.4=py310h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pyopengl=3.1.6=pypi_0
pyopenssl=23.0.0=py310h06a4308_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.10=h7a1cb2a_2
pyyaml=6.0=py310h5eee18b_1
readline=8.2=h5eee18b_0
requests=2.29.0=py310h06a4308_0
ruamel.yaml=0.17.21=py310h5eee18b_0
ruamel.yaml.clib=0.2.6=py310h5eee18b_1
sentry-sdk=1.9.0=py310h06a4308_0
setproctitle=1.2.2=py310h7f8727e_0
setuptools=65.6.3=py310h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
sqlite=3.41.2=h5eee18b_0
tbb=2021.8.0=hdb19cb5_0
tk=8.6.12=h1ccaba5_0
toolz=0.12.0=py310h06a4308_0
torch=2.0.1=pypi_0
tqdm=4.65.0=py310h2f386ee_0
typing_extensions=4.5.0=py310h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py310h06a4308_0
wandb=0.15.2=pyhd8ed1ab_0
wheel=0.38.4=py310h06a4308_0
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py310h5eee18b_0
